[hw2] added some final touches to CRUD and basic APIs

+++ src/main/java/olympic/Solution.java
@@ -115,11 +115,11 @@ public class Solution {
         try {
 
             /*=================ShareSport View=============================== */
-            pstmt = connection.prepareStatement("CREATE View ShareSport(athlete1_id, athlete2_id, sport_id, athlete1_took, athlete2_took) as" +
-                    "SELECT a1.athlete_id  , a2.athlete_id , s.sport_id," +
-                    "(SELECT count(*) FROM Goes_to as g1 WHERE g1.athlete_id = a1.athlete_id and s.sport_id = g1.sport_id)," +
-                    "(SELECT count(*) FROM Goes_to as g2 WHERE g2.athlete_id = a2.athlete_id and s.sport_id = g2.sport_id)" +
-                    "FROM Athlete a1, Athlete a2, Sports s");
+            pstmt = connection.prepareStatement("CREATE View ShareSport(athlete1_id, athlete2_id, sport_id, athlete1_took, athlete2_took) as " +
+                    "SELECT a1.athlete_id  , a2.athlete_id , s.sport_id, " +
+                    "(SELECT count(*) FROM Goes_to as g1 WHERE g1.athlete_id = a1.athlete_id and s.sport_id = g1.sport_id), " +
+                    "(SELECT count(*) FROM Goes_to as g2 WHERE g2.athlete_id = a2.athlete_id and s.sport_id = g2.sport_id) " +
+                    "FROM Athlete a1, Athlete a2, Sports s;");
             pstmt.execute();
         } catch (SQLException e) {
             //e.printStackTrace()();
@@ -273,6 +273,18 @@ public class Solution {
             }
             return ERROR;
         }
+        finally {
+            try {
+                pstmt.close();
+            } catch (SQLException e) {
+                //e.printStackTrace()();
+            }
+            try {
+                connection.close();
+            } catch (SQLException e) {
+                //e.printStackTrace()();
+            }
+        }
 
         return OK;
     }
@@ -298,22 +310,29 @@ public class Solution {
         } catch (SQLException e) {
             return Athlete.badAthlete();
         }
+        finally {
+            try {
+                pstmt.close();
+            } catch (SQLException e) {
+                //e.printStackTrace()();
+            }
+            try {
+                connection.close();
+            } catch (SQLException e) {
+                //e.printStackTrace()();
+            }
+        }
+
     }
 
     public static ReturnValue deleteAthlete(Athlete athlete) {
         Connection connection = DBConnector.getConnection();
         PreparedStatement pstmt = null;
-        int tmp = 0;
+        Athlete a = getAthleteProfile(athlete.getId());
+        if (a.getId() == -1) {
+            return NOT_EXISTS;
+        }
         try {
-            pstmt = connection.prepareStatement(
-                    "SELECT COUNT(*) " +
-                            "FROM Athlete \n" +
-                            "WHERE athlete_id = ?;");
-            pstmt.setInt(1, athlete.getId());
-            ResultSet results = pstmt.executeQuery();
-            while (results.next()) {
-                tmp = results.getInt("COUNT");
-            }
             pstmt = connection.prepareStatement(
                     "DELETE FROM Athlete " +
                             "WHERE athlete_id = ? ;");
@@ -333,9 +352,6 @@ public class Solution {
             } catch (SQLException e) {
                 //e.printStackTrace()();
             }
-        }
-        if (tmp == 0) {
-            return NOT_EXISTS;
         }
         return OK;
     }
@@ -364,6 +380,18 @@ public class Solution {
             }
             return ERROR;
         }
+        finally {
+            try {
+                pstmt.close();
+            } catch (SQLException e) {
+                //e.printStackTrace()();
+            }
+            try {
+                connection.close();
+            } catch (SQLException e) {
+                //e.printStackTrace()();
+            }
+        }
 
         return OK;
     }
@@ -389,22 +417,28 @@ public class Solution {
         } catch (SQLException e) {
             return Sport.badSport();
         }
+        finally {
+            try {
+                pstmt.close();
+            } catch (SQLException e) {
+                //e.printStackTrace()();
+            }
+            try {
+                connection.close();
+            } catch (SQLException e) {
+                //e.printStackTrace()();
+            }
+        }
     }
 
     public static ReturnValue deleteSport(Sport sport) {
         Connection connection = DBConnector.getConnection();
         PreparedStatement pstmt = null;
-        int tmp = 0;
+        Sport s = getSport(sport.getId());
+        if (s.getId() == -1) {
+            return NOT_EXISTS;
+        }
         try {
-            pstmt = connection.prepareStatement(
-                    "SELECT COUNT(*) " +
-                            "FROM Sport \n" +
-                            "WHERE sport_id = ?;");
-            pstmt.setInt(1, sport.getId());
-            ResultSet results = pstmt.executeQuery();
-            while (results.next()) {
-                tmp = results.getInt("COUNT");
-            }
             pstmt = connection.prepareStatement(
                     "DELETE FROM Sport " +
                             "WHERE sport_id = ? ;");
@@ -425,9 +459,6 @@ public class Solution {
                 //e.printStackTrace()();
             }
         }
-        if (tmp == 0) {
-            return NOT_EXISTS;
-        }
         return OK;
     }
 
@@ -442,14 +473,14 @@ public class Solution {
         PreparedStatement pstmt = null;
         try {
             // check if athlete already joined the sport
-            pstmt = connection.prepareStatement("SELECT * FROM Goes_to " +
-                    "WHERE sport_id=? AND athlete_id=?;");
-            pstmt.setInt(1, sportId);
-            pstmt.setInt(2, athleteId);
-            ResultSet results = pstmt.executeQuery();
-            if (results.next()) {
-                return ALREADY_EXISTS;
-            }
+//            pstmt = connection.prepareStatement("SELECT * FROM Goes_to " +
+//                    "WHERE sport_id=? AND athlete_id=?;");
+//            pstmt.setInt(1, sportId);
+//            pstmt.setInt(2, athleteId);
+//            ResultSet results = pstmt.executeQuery();
+//            if (results.next()) {
+//                return ALREADY_EXISTS;
+//            }
 
             // if athlete is active - update athletes’ counter, add hes fee
             int fee = 100;
@@ -471,6 +502,9 @@ public class Solution {
             pstmt.execute();
         } catch (SQLException e) {
             //e.printStackTrace()();
+            if (Integer.valueOf(e.getSQLState()) == PostgreSQLErrorCodes.UNIQUE_VIOLATION.getValue()) {
+                return ALREADY_EXISTS;
+            }
             return ERROR;
         } finally {
             try {
@@ -610,12 +644,6 @@ public class Solution {
             if (!results.next()){
                 ret = NOT_EXISTS;
             }
-            else{
-                int iVal = results.getInt("place");
-                if (results.wasNull()) {
-                    ret = NOT_EXISTS;
-                }
-            }
 
             // disqualifying athlete - nullifying their medal
             pstmt = connection.prepareStatement("UPDATE Goes_to " +
@@ -652,17 +680,17 @@ public class Solution {
         Connection connection = DBConnector.getConnection();
         PreparedStatement pstmt = null;
         try {
-            // check if athletes are already friends
-            pstmt = connection.prepareStatement("SELECT * FROM Friends " +
-                    "WHERE (Athlete1_ID=? AND Athlete2_ID=?) OR (Athlete1_ID=? AND Athlete2_ID=?);");
-            pstmt.setInt(1, athleteId1);
-            pstmt.setInt(2, athleteId2);
-            pstmt.setInt(3, athleteId2);
-            pstmt.setInt(4, athleteId1);
-            ResultSet results = pstmt.executeQuery();
-            if (results.next()) {
-                return ALREADY_EXISTS;
-            }
+//            // check if athletes are already friends
+//            pstmt = connection.prepareStatement("SELECT * FROM Friends " +
+//                    "WHERE (Athlete1_ID=? AND Athlete2_ID=?) OR (Athlete1_ID=? AND Athlete2_ID=?);");
+//            pstmt.setInt(1, athleteId1);
+//            pstmt.setInt(2, athleteId2);
+//            pstmt.setInt(3, athleteId2);
+//            pstmt.setInt(4, athleteId1);
+//            ResultSet results = pstmt.executeQuery();
+//            if (results.next()) {
+//                return ALREADY_EXISTS;
+//            }
 
             // create friendship
             pstmt = connection.prepareStatement("INSERT INTO Friends " +
@@ -678,6 +706,10 @@ public class Solution {
             if (Integer.valueOf(e.getSQLState()) == PostgreSQLErrorCodes.CHECK_VIOLATION.getValue()) {
                 return BAD_PARAMS;
             }
+            if (Integer.valueOf(e.getSQLState()) == PostgreSQLErrorCodes.UNIQUE_VIOLATION.getValue()) {
+                return ALREADY_EXISTS;
+            }
+
             return ERROR;
         } finally {
             try {
@@ -857,7 +889,7 @@ public class Solution {
                 number_of_medals = results.getInt("COUNT");
             }
         } catch (SQLException e) {
-            return number_of_medals;
+            return 0;
         }
         finally {
             try {
@@ -889,7 +921,7 @@ public class Solution {
                 income = results.getInt("SUM");
             }
         } catch (SQLException e) {
-            return income;
+            return 0;
         }
         finally {
             try {
@@ -924,7 +956,7 @@ public class Solution {
                 best_country = results.getString("country");
             }
         } catch (SQLException e) {
-            return best_country;
+            return "";
         }
         finally {
             try {
@@ -949,7 +981,6 @@ public class Solution {
             pstmt = connection.prepareStatement(
                     "SELECT city " +
                             "FROM Sport as s " +
-                            "LEFT JOIN Goes_to as g ON g.Sport_id = s.Sport_id " +
                             "GROUP BY city HAVING COUNT(distinct s.Sport_id)>0 " +
                             "ORDER BY SUM(athlete_count)*1.0/COUNT(distinct s.Sport_id) DESC, s.city desc;" );
             ResultSet results = pstmt.executeQuery();
